[DO NOT MERGE] configure the PGO bot --with-readline=edit on 3.10+ builds
This is just to add coverage of that configure option to at least one buildbot. I control that one so it makes sense.

Currently blocked by https://bugs.python.org/issue43172 being fixed.  I'm leaving the edit as a proposed change for now so I don't lose track of it.