gh-419: Allow unstable builders to trigger on PRs
This intention of this PR is to make it possible to *manually trigger* also unstable buildbots on PRs using the `!buildbot` command (and labels).

The lists of "PR builders" are used in several places where I kept it to *only* stable PRs, because those other places didn't seem related to the manual triggering of buildbots on PRs.