Increase test_timeout on RHEL8-ppc64le for older branches
This particular worker is slow: a successful test_gdb run usually
will take about 17 minutes, but often it fails on the 20-minute
timeout.

On 3.11+, the tests are split up into three 7-min suites so this
isn't a problem, but for older branches it's probably best to
solve it in the buildbot config.