Decrease parallel tests on itamaro ubuntu worker (#488)
This worker was having a hard time recently.
I think it might be that now the builders it's running are covering more branches (3.13 + main), so it is more likely now that multiple builders are running concurrently. I want to try reducing parallel tests (4 is more in line with the typical configuration) to see if it improves things (if it doesn't, we can try reducing parallel builders).