Fix the dashboard (#609)
Always yield a BuildFailure for failing builders.

BuildFailure now allows the latest & breaking build to be the same,
and deduplicates in that case.

Also, show builders without detected problems
Only 3 builds are shown for these.

Also, reduce the number of builds shown from 70 to 50.
Add an ellipsis if some are hidden.